rename the main folder
